moving SDL specific code to sdl_canvas.go and removing Get from getter names: https://golang.org/doc/effective_go.html#Getters
