+++ examples/sine-wave/main.go
@@ -25,7 +25,7 @@ type sineWaveDrawable struct {
 }
 
 func (s *sineWaveDrawable) Init(canvas eff.Canvas) {
-	numPoints := (cols-1)*canvas.GetHeight() + (rows-1)*canvas.GetWidth()
+	numPoints := (cols-1)*canvas.Height() + (rows-1)*canvas.Width()
 	s.tx = math.Pi / 9
 	s.ty = math.Pi / 4
 	s.xFreq = 1
@@ -36,12 +36,12 @@ func (s *sineWaveDrawable) Init(canvas eff.Canvas) {
 	s.gridPoints = make([]eff.Point, numPoints)
 	s.origGridPoints = make([]eff.Point, numPoints)
 	index := 0
-	cellWidth := math.Ceil(float64(canvas.GetWidth()) / float64(cols))
-	cellHeight := math.Ceil(float64(canvas.GetHeight()) / float64(rows))
+	cellWidth := math.Ceil(float64(canvas.Width()) / float64(cols))
+	cellHeight := math.Ceil(float64(canvas.Height()) / float64(rows))
 	// Create Columns
 	for i := 1; i < cols-1; i++ {
 		x := i * int(cellWidth)
-		for j := 0; j < canvas.GetHeight(); j++ {
+		for j := 0; j < canvas.Height(); j++ {
 			s.gridPoints[index] = eff.Point{X: (x), Y: (j)}
 			s.origGridPoints[index] = eff.Point{X: (x), Y: (j)}
 			index++
@@ -51,7 +51,7 @@ func (s *sineWaveDrawable) Init(canvas eff.Canvas) {
 	// Create Rows
 	for i := 1; i < rows-1; i++ {
 		y := i * int(cellHeight)
-		for j := 0; j < canvas.GetWidth(); j++ {
+		for j := 0; j < canvas.Width(); j++ {
 			s.gridPoints[index] = eff.Point{X: (j), Y: (y)}
 			s.origGridPoints[index] = eff.Point{X: (j), Y: (y)}
 			index++
@@ -87,7 +87,7 @@ func (s *sineWaveDrawable) Update(canvas eff.Canvas) {
 	}
 
 	for i, point := range s.origGridPoints {
-		newX, newY := sineWaveDistortXY(point.X, point.Y, canvas.GetWidth(), canvas.GetHeight())
+		newX, newY := sineWaveDistortXY(point.X, point.Y, canvas.Width(), canvas.Height())
 		s.gridPoints[i] = eff.Point{X: newX, Y: newY}
 	}
